Major cleanup. Use instanced renderer only for symbol and rect marks
