+++ avenger-wgpu/src/marks/mod.rs
@@ -1,16 +1,10 @@
-pub mod arc;
-pub mod basic_mark;
-pub mod gradient;
-pub mod image;
 pub mod instanced_mark;
 pub mod multi;
-pub mod path;
 pub mod rect;
-pub mod rule;
 pub mod symbol;
 
-pub mod gradient2;
-pub mod image2;
+pub mod gradient;
+pub mod image;
 
 #[cfg(feature = "text-glyphon")]
 pub mod text;
